美食服务订单管理

+++ czb-bn/bn-service-impl/src/main/java/com/cqut/czb/bn/service/impl/food/ManageSystemFoodServiceImpl.java
@@ -176,9 +176,13 @@ public class ManageSystemFoodServiceImpl implements ManageSystemFoodService{
            }
         }
 
-        mapper.deleteList(deleteList);
+        if(deleteList.size() != 0) {
+            mapper.deleteList(deleteList);
+        }
 
-        mapper.insertSets(insertList);
+        if(insertList.size() != 0) {
+            mapper.insertSets(insertList);
+        }
         return true;
     }
 }
